Fix stream reset error in blobs_sidecars_by_range
We previously assumed that a db query for blob sidecars could return nil
if there are no blobs associated with a block. This assumption no longer
holds as we insert "zero-valued" blob sidecars into the db for
every beacon block. As such, the RPC handler unnecessarily implement throttling
for every sidecar message, causing timeouts and the client to reset
the stream.